Fixed scroll bar bug
scroll bar was scrolling outside of the container
This was due to the handle scroll function not taking into account
the margin on the sides of the scroll bar container.
Fixed by adding a scroll bar track div within the scroll bar container.
This allowed for use of padding instead of margin which works with the
handle scroll function