rework of mouse move function for y coordinate
the cursor container element was changed to position relative in order to
facilitate hiding overflow. The find offset function in the cursor wrapper
component was bugged as an element with an offset of zero would return
out of the recursion early. Y is now calculated from bounding client rects top on currsor outer as well as the offset from the container width
divided by 2