Bug fix, incorrect cursor position on entry before cursor reset
if the user enters the container before the cursor is reset upon leaving
the cursor would be misaligned. The tranform property is set to 0
on leave so the coordinate positions on entry needed to account for
the difference between where the cursor is on entry and the starting
position. cursorAdjustment state was created for this fix