rework of cursor move functions
instead of using page x and page y to calculate changes in the cursor
position client x and client y are used in relation to the current
position of the cursor. The inital position of the cursor is set on
load and is not reset unless the page is scrolled.