cursor takeover now uses cursor state for managing movement of cursor
previously a ref was used to store the positional values of the cursor
and an additional state was used to trigger rerenders. Now a useCallback
with a cursor state dependency is used to track changes in cursor.
This allows all the information needed for cursor changes to be in
one location.