Bug fix: cursor misalignment on scroll
When the window was scrolled after leaving the cursor container and
whilst the cursor was transitioning back to its starting position,
reentry would cause misalignment. The cause was the setCursorPosition
function taking the incorrect positions of the cursor during the
transition. To solve an extra condition was added to scroll handler
to check if the cursor is in the transition. If it is then the cursor
position is calculated with the starting position and the scrolldiff